+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/utils/extensions/Throwable.kt
@@ -1,0 +1,10 @@
+package com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions
+
+import java.io.ByteArrayOutputStream
+import java.io.PrintStream
+
+fun Throwable.collectStackTrace(): String {
+    return ByteArrayOutputStream().also {
+        printStackTrace(PrintStream(it))
+    }.toString(Charsets.UTF_8.toString())
+}