+++ CHANGELOG.md
@@ -39,3 +39,8 @@ recommended to register/unregister in this table
 * Rename methods of `PostsLikesMessagesTable` to be consistent with their
 behaviours
 * ***Now plugins `HAVE NO VERSIONS` and must not implement `onInit` method***
+
+## 0.3.9
+
+* Add `Throwable#collectStackTrace` which return as string stacktrace of
+caller